cgroups: Fix wrong permissions on mock CGroup directories
We use wrong permission to create the mock CGroup hierarchy such that
the directories do not have the executable bit set. This works by chance
given that we (invalidly) create the mock hierarchy after the test has
done its logic already, and subsequently the cgroups package has already
created the hierarchy for us. The call to `os.MkdirAll()` is thus
effectively a no-op waiting to do the wrong thing if it does something.

Fix this by using 755 permissions.