cgroups: Fix mock hierarchy being created after logic was executed
To test CGroups, we create a mock hierarchy by creating the expected
files ourselves in a temporary directory. The only callsite which does
set up the hierarchy does so too late though, and as a result the
hierarchy was created for us by the cgroups package already. So while
the cgroups package knows to create these files itself if the hierarchy
is missing, it will in fact create those files with no permission bits
being set at all. This is not quite a bug in the upstream package: these
files would always exist in a "real" CGroups hierarchy.

Fix the order such that we first create the CGroups hierarchy ourselves.
This gives us the chance to create those files with the correct
permission bits.