glsql: PingContext doesn't hang and respects context
lib/pq driver has a problem with establishment of the
connection to the incorrect address that accepts TCP
connections. If it tries to connect not to the PostgreSQL
database in hangs forever and it doesn't matter if
provided context is cancelled in some point in time
or not. The jackc/pgx driver has no such a problem.
That is why in this change we remove dirty hack used
to not hang on service startup.