repository: Unconditionall enable atomic repo creation semantics
In 896612200 (Merge branch 'pks-enable-atomic-rename-repository' into
'master', 2021-12-13), we have implemented new locking semantics for
RPCs which create or move repositories. Like this, repositories which
are about to be modified will first be locked such that no concurrent
request can try to rename, recreate or remove the repository at the same
point in time. Furthermore, repositories are created in a temporary
directory first, which allows us to easily discard the repository in
case any error happens during creation of the repository.

The feature flag guarding the code has been enabled on December 6th
without any issues so far. Remove the featureflag to always enable the
code.

Changelog: changed