repository: Unconditionall enable atomic RemoveRepository semantics
In f328e3f70 (repository: Implement atomic locking semantics for
RemoveRepository, 2021-11-30), we have implemented new locking semantics
for the RemoveRepository RPC. Like this, repositories which are about to
be removed will first be locked such that no concurrent request can try
to rename, recreate or remove the repository at the same point in time.
Furthermore, this changes semantics such that the RPC now returns a
NotFound error in case the repository doesn't exist.

The feature flag guarding the code has been enabled on December 15th
without any issues so far. Remove the featureflag to always enable the
code.

Changelog: changed