commit: Refactor FindAllCommits test to not modify global function
The FindAllCommits RPC is currently using a global variable which keeps
track of the function that's used to list local branches in the repo.
This is done such that tests can switch this function to a different one
and thus control which local branches are found in the repository. This
is quite roundabout though: instead of controlling the function, we can
just modify the repository to have the required state.

Convert the test to modify the repository instead of the function.