commit: Convert FindAllCommits to use localrepo abstraction
The FindAllCommits RPC is using `ref.FindBranchNames` to identify local
branches, which is assigned to a local variable. This was done like this
because one of the tests swapped the function with a different one, but
this test has since been refactored to not modify global state anymore.

Convert the RPC to use the localrepo abstraction to find local branches
and get rid both of the global variable and the new unused function
`ref.FindBranchNames`.