UploadPack: use 64KB copy buffer
This increases the copy buffer size for PostUploadPackWithSidechannel,
PostUploadPack and SSHUploadPack. The old value was the Go stdlib
default (32KB). The new value is 64KB. Because we patched
git-upload-pack to do (roughly) 64KB writes to its stdout, it is
optimal to also do 64KB reads in Gitaly. This reduces the number of
system calls needed to proxy git-upload-pack output to the client.

Changelog: performance