git: Remove Git binary path from the hooks payload
Originally, we used to execute Git in the pack-objects-hook, which
required us to know about the Git executable's path in gitaly-hooks.
This has been removed in commit f069a4680 (Add pass-through
PackObjectsHook RPC, 2021-02-11), which instead invoked an RPC and thus
moved the responsibility into the main Gitaly server process. The binary
path we inject into gitaly-hooks is since unused.

Remove it.