gittest: Convert HTTP server to use command factory
The test Git HTTP server is currently constructing the Git environment
from the configuration, which is deprecated and will eventually stop to
work. Convert it to get a Git command factory as argument instead.

While at it, this also renames the function from `GitServer` to
`HTTPServer`, which more clearly indicates what it's about.