gittest: Implement new intercepting command factory
We have several tests which intercept Git commands. This is done by
replacing the Git binary path with a custom one pointing to a script,
This pattern will cause problems when we migrate to feature-flagged Git
binary paths via the Git command factory.

Introduce a new InterceptingCommandFactory which allows tests to satisfy
this usecase.