git: Drop useless check in CGroup tests
The command factory's CGroup tests verify that the Git path is indeed
set after calling `SetGitPath()`. This is not the subject of this test,
so we should assume that it's been tested somewhere else and works as
expected. Furthermore, testing the binary path directly like this will
eventually stop working.