git: Allow command factories to set up state
Creating Git command factories cannot currently error, which means that
it cannot set up any state which may fail to be created. We're going to
move construction of the hooks environment into the Git command factory
though, and that'll require us to create this environment on the
filesystem.

Refactor `NewExecCommandFactory()` to be able to return both a cleanup
function and an error. While these aren't yet used, they'll allow us to
eventually construct more state for the factory.