git: Remove unused binary directory from hooks payload
The Gitaly binary directory endcoded into the hooks payload has never
actually been used. Originally, the intent was for it to replace the
GITALY_BIN_DIR environment variable, but that doesn't work given that it
is required by our gitlab-shell-hook script to locate the gitaly-hooks
binary, and we don't want to require a JSON-decoder there.

Remove the variable.