backup: Convert tests to use `StartPraefect()`
Some of the backup tests verify that backup and restoration works as
expected with Praefect. The test setup manually bootstraps the Praefect
binary, which is non-trivial, and later it has to send multiple requests
to create a repository until Praefect has started up correctly.

Convert it to use `testserver.StartPraefect()`, which runs the Praefect
server and waits for it to be healthy.