updateref: Drop unused Gitaly configuration
The updateref package used to require the Gitaly configuration given
that it needed to be passed when constructing Git command options. Now
that such dependencies are instead provided by the Git command factory
itself the configuration is not used anymore by this package.

Remove the dependency on the Gitaly configuration and adjust callers to
not pass it anymore.