localrepo: Let setup testhelper accept options
The setup testhelper in the localrepo package currently accepts a single
boolean option, which controls whether it'll create an empty repository
or clone the test repo. We're about to add another boolean option
though, and having to pass two unnamed booleans is an awkward calling
convention given that one always needs to guess what their meaning is.

Refactor the function to instead accept options. While at it, the
function is moved into "testhelper_test.go" given that it's a general
helper function that's reused across different files.