localrepo: Add option to disable hooks in testhelper
Currently, we disable hooks in localrepo tests by modifying the Gitaly
configuration _after_ we have already constructed the Git command
factory. This works right now given that the config is passed at the
callsite when spawning new commands, but it will stop working when this
is changed such that the Git command factory uses its own configuration
given that we modify the config too late. As a result, the Git command
factory would run without its hooks being disabled.

Add a new option to the `setupRepo()` test helper which disables hooks
for us. This'll allows us to move setup of hooks into `setupRepo()` such
that we can modify the Gitaly configuration before constructing the Git
command factory.