git: Move hook verification into the command factory
The Git command factory is the new home of the hooks directory.
Furthermore, the way hooks are set up and thus the way we'll eventually
have to verify they work as expected will change when we migrate from
the Ruby hooks directory to a hooks directory that's created at runtime
in the Git command factory.

Prepare our codebase for this change by moving verification of Git hooks
from the Gitaly config code to the Git command factory.