cmd/gitaly-hooks: Refactor Gitaly config test setup
Tests for the "gitaly-hooks" command set up a Gitaly configuration by
manually writing a formatted string into a file. This is much more
complex than it needs to be given that we can just serialize a real
Gitaly configuration and write it there, and furthermore the resulting
configuration is incomplete. This hasn't been a problem until now, but
it's going to become one soon when the Git command factory starts to
verify that Git hooks are set up as expected, which needs the path to
the Ruby directory to work correctly.

Refactor the test setup to always serialize a complete, valid
configuration instead of crafting a custom one.