cmd/gitaly: Use option to skip hooks
We're listening for a global variable "GITALY_TESTING_NO_GIT_HOOKS" in
order to not set up hooks at all. This is used by Rails' test suite
because they do not have a complete Gitaly setup available. The way this
is implemented is that we always check whether this environment variable
is set whenever executing a Git command. As a result, we always depend
on global state here.

Refactor the code to instead only check this environment variable when
starting up the Gitaly server. If set, then we pass the `SkipHooks()`
option to the Git command factory. Like this, "normal" code doesn't
depend on the process's environment anymore.