cmd/gitaly-wrapper: Add tests for finding processes by PID file
Now that `readPIDFile()` doesn't depend on global state anymore, it's
also a lot easier to test `findProcess()` via a PID file given that it
also doesn't depend on environment variables anymore, but instead
accepts a parameter. Add a bunch of tests to verify it works as
expected.

While at it, rename the function to `findProcess()` instead of
`findGitaly()`. The executable is used for both Praefect and Gitaly, so
the found process may not in fact be a Gitaly instance.