cmd/gitaly-wrapper: Add tests for `isGitaly()`
When gitaly-wrapper tries to adopt a process, then it first verifies
that it is in fact the expected process. This is to guard against PID
wraparound, where the existing PID file contains a PID which has since
been taken by a completely unrelated command.

Add tests to verify this function works as expected. While at it, rename
it to `isExpectedProcess()`: the wrapper is not only used for Gitaly,
but also for Praefect and may thus not in fact check for Gitaly.