cmd/gitaly-wrapper: Rename function to spawn binary
The function to spawn a new wrapped binary is currently called
`spawnGitaly()`, which stems from a time where it was really only used
for Gitaly. Nowadays it's also used for Praefect though, so this name
doesn't make a lot of sense anymore.

Rename it to `spawnProcess()` to avoid any confusion.