gittest: Provide helper function to create command factories
We're about to change the ExecCommandFactory to carry state soonish,
which both requires us to return an error code and a cleanup function.
This is going to become tedious to manage in our tests given that
there's many locations where we construct the factory.

Introduce a new function `gittest.NewCommandFactory()` to be used in
tests, such that callsites don't have to handle all of this.