git: Move computing the hooks path into the Git command factory
Right now, the hooks path always points into Gitaly's Ruby directory,
which contains the hooks location due to legacy reasons. Given that
we will some day phase out our dependency on Ruby, we also have to
migrate the hooks directory somewhere else.

This "somewhere else" could easily be a different on-disk directory. But
using an on-disk directory has proven to be one more breaking point in
the past, where we've had multiple issue reports where customers
unknowingly replaced hooks stored in there with their own ones. This
naturally doesn't work, and customers must instead use the custom hooks
mechanism to install any custom server-side hooks. Furthermore, having
an on-disk location requires us to intercept hook executions with a
wrapper script which is able to locate the gitaly-hooks binary via an
injected environment variable.

We can do better though: instead of having an on-disk directory, we can
just create the hooks path at runtime. That is, when we start up Gitaly,
we can create a temporary directory and then create the necessary
symlinks which directly point to the "gitaly-hooks" binary. This removes
the chance to break the setup by mistake, and furthermore it lets us get
rid of the wrapper script.

This will require us to create some state though, which isn't really
well-located in the config code which currently computes the hooks path.
A much better location would be the Git command factory, which can set
up the hooks directory at creation-time and then return the correct path
to callers.

Move the hooks path implementation from the configuration code into the
Git command factory as a first step towards this architecture.