datastore: Fix test with Postgres listening on Unix socket
The test which creates a new listener with an invalid SSL mode is
failing on machines where Postgres is listening on a Unix socket instead
of using TCP. This is because when using Unix sockets, pgx will not even
parse the SSL mode and thus not notice that it contains an invalid mode.

Fix this by amending the database configuration to specify a TCP-based
host so that pgx always parses the SSL mode.