gittest: Execute commands via Git command factory's environment
We're about to move constructing of the Git execution environment ino
the Git command factory, which means that the Gitaly configuration will
not always carry a valid path to the Git binary anymore. The gittest
helpers which execute Git commands rely on that to be the case though.

Prepare for this change by just creating an ad-hoc Git command factory,
which will then set up the execution environment for us.