command: Fix log message verification
The root cause of the change is a flaky behaviour
of a couple of the tests in the command package.
The tests assume the first log message would be
the output of the bash script that is invoked.
But it happens that 'spawn token acquired' text
is the first message instead. The output depends
on the execution time as this message appears only
if token acquire operation takes more than 5 ms.
In any case the message from the script would be the
last in the log output that is why extractMessage
renamed to extractLastMessage and now returns the
last message found.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3980