praefect: Fix TestReplMgr_ProcessStale
Calling the last ticker.Tick() puts the "tick"
into the channel in unblocking way as the channel
is buffered by 1. It happens the cancel() is called
before "tick" is consumed by the ProcessStale method.
So the select used inside the ProcessStale consumes
observes cancelled context and finishes the execution.
By moving invocation of the cancel() into the handling
method we are sure the select loop will consume all
"tick"s first and context cancellation will be only
option on the next loop iteration.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3993