ci: Cache dependencies in "verify" job
The "verify" CI job executes the "verify" Makefile target, which in turn
verifies that protobuf definitions are up-to-date. As a consequence, it
needs the protoc compiler, but we're not using the dependencies cache
for this target.

Add the dependencies cache to the job to avoid rebuilding of protoc
every time the job runs and make sure that protoc is built.