Makefile: Run tests with bundled and external Git configured
We currently have two test flavors, one which runs Gitaly tests with
bundled Git and one which runs them with external Git. This doesn't
allow us to test the new feature-flag to toggle between both, given that
we'd need to have both configured.

Change our tests to set up both in case WITH_BUNDLED_GIT is unset to
test the feature flag. Note that this is preferable over injecting both
in case WITH_BUNDLED_GIT is set: CI is set up to delete the Git install
directory in that case to verify bundled Git can execute as a standalone
binary, and we'll want to retain this sanity check.