featureflag: Allow injecting outgoing and incoming flags with value
The function `ContextWithFeatureFlags()` will inject a set of feature
flags as `true` into both the outgoing and incoming context, which is
used in our tests. It's not possible to use this helper to inject a
disabled flag though, and we're about to need that.

Refactor the helper to only accept a single flag and a value. This also
brings it in line with the other helpers which inject feature flags, and
there aren't any callsites which use it to inject multiple flags anyway.