Makefile: Allow skipping "build" dependency for rspec target
We're about to switch over our CI build to run tests as an unprivileged
user. This means that we cannot write anything into the source directory
at all because we ain't go the permission to do so. The "rspec" target
currently depends on the "build" target though, most imporantly to get
the gitaly-hooks binary, and because it writes into the source directory
the build now fails.

Add a workaround by conditionally disabling the "build" dependency for
the "rspec" target. The CI job will just manually build the binaries
before running "rspec" and then set the flag.

While at it, remove the "prepare-tests" dependency from "test-ruby":
it's depending on "rspec" already, which transitively depends on
"prepare-tests".