git: Add option to override the Git binary path
The intercepting Git command factory needs to override the path to the
Git binary such that it can be intercepted with a custom script used for
testing purposes. This used to work alright until now given that we
could just ignore the Git binary path in the Gitaly configuration and
disable the use of bundled Git. But that's about to change given that
we'll always set up execution environments for both bundled and external
Git and choose one or the other via a feature flag.

Prepare for this change by adding a new option to the Git command
factory which lets callers construct it with an overridden Git binary
path.