git: Set up all configured execution environments
We support two different execution environments with the bundled and
external Git configurations. Right now, we only support either one or
the other to be configured by the user. This has the downside that
migrating to bundled Git is an all-or-nothing action, which means it
carries quite a lot of risk.

Prepare for the introduction of a feature flag by always setting up both
execution environments. Until we have that feature flag, bundled Git
will always be chosen over external Git.