ref: Removal of unused all_refs parameter
The request parameter all_refs was never used by
the gitaly. It is not used by the call sites as well
so we remove it from the code. The getReplicationDetails
function now returns a hardcoded false value for
that parameter to support backwards compatibility.
It should return nil on the next release (14.8).

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3997
Changelog: deprecated