git: Move Git version detection into command factory
The code to detect the Git version is using the Git command factory to
do so. Given that it's not a receiver function it thus cannot easily
reuse state, and as a result we cannot cache Git version information.

Move version detection into the Git command factory itself to prepare
for caching it.