git: Cache Git version
In various places we need to detect the Git version such that we can
decide whether a given feature is supported. Detecting the Git version
requires us to run git-version(1) every time, which can become quite
expensive when executed often. This is exacerbated under load when
spawning lots of commands due to the spawn tokens we generate, which
means that it may take multiple seconds to detect the Git version.

We can improve the situation quite easily by caching the Git version: if
the Git binary has not been changed since we have last detected the Git
version, then we don't need to rerun git-version(1). This reduces the
number of Git commands we need to spawn in some locations and thus also
reduces the contention around command spawn tokens.

Changelog: performance