git: Move hook directory override into the command factory
Tests have the ability to override the hook directory by setting a field
in the Gitaly configuration which isn't being serialized or deserialized
and thus only for internal use. Now that all hook-directory-related code
is hosted in the Git command factory it's more sensible to instead
implement this via an option that can be passed to the command factory
at creation time.

Introduce a new `WithHooksPath()` option and convert all callers to use
it, allowing us to remove the old hooks override in the config.