git: Prepare for the setup of multiple hook directories
We're about to move away Gitaly's hooks directory from the Ruby
directory into a temporary directory. Given that this is quite a big
change and thus risky we want to do this via a feature flag, which thus
requires us to set up both old and new hooks directories.

Refactor the code to allow for the setup of multiple hook directories.