git: Rewrite version tests to use intercepting command factory
The tests which verify that we correctly compute and cache the Git
version via the command factory are currently constructing a real Git
command factory whose Git binary path is replaced. This is going to stop
working soonish with the introduction of feature flags to switch between
Git versions. Instead, all commands which want to replace Git with their
own binary shall use the intercepting Git command factory provided by
the gittest package.

Rewrite tests to do so. Note that we also have to change semantics of
the intercepting command factory to use intercepting script to compute
the Git version instead of using the real Git executable, or otherwise
we wouldn't be use our own script to return arbitrary version strings.
No callsite which uses the intercepting command factory depends on the
Git version right now though, so this is fine for the time being.