git: Cache Git version by binary path
We're about to introduce support for multiple different Git execution
environments, which means that we can in theory also have multiple
different Git versions. This breaks the current way we do caching, where
we only store a single cached Git version and reuse. If we alternate the
binary path though via the use of a feature flag, then we'd repeatedly
invalidate the cached version because of the stat information differing
between versions.

Improve the cache by storing cached Git versions by binary path.