git: Convert UploadPackEnv into a Git command option
The `UploadPackEnv()` helper function is used to set up a Git command
such that it can fetch changes from another internal Gitaly server. It
requires the caller to inject the Gitaly configuration to do so such
that it can derive the binary directory where the `gitaly-ssh` binary is
located, which in turn requires services to carry a Gitaly config, as
well.

Convert the helper into a Git command option, which allows us to take
the Gitaly configuration from the factory instead. This prepares for the
removal of the Gitaly configuration from most Gitaly services.