streamcache: Move metric tracking whether cache is enabled or not
We're exposing a metric which reports whether the pack-objects cache is
enabled or not. This metric is set when creating the hook service, which
is a bit roundabout: we can instead deduplicate instances of the stream
cache and then report whether it is enabled or not in its constructor.
This removes knowledge about whether the cache is enabled or not from
the hook service and furthermore removes its dependency on the Gitaly
configuration.