services: Drop dependency on the Gitaly configuration
We have recently noticed that the way we handle the Gitaly configuration
it can still be considered to be a pseudo-global variable. It's passed
down into most of our components, which makes it hard to reason about
the impact the configuration is having on a global level given that e.g.
fields like the Git binary path were accessed everywhere. The result is
that it's hard to tweak the Gitaly configuration and change its
behaviour.

A more modern way of handling this is to never pass down the Gitaly
configuration into our services. Instead, it should only be used to
construct a set of dependencies like e.g. the Git command factory and
then inject those dependencies into our services. This decouples the
services from the configuration, and furthermore it encapsulates use of
the configuration into those smaller parts, which makes it easier to
reason about.

In fact, most of our services don't depend on the Gitaly configuration
at all anymore now that most of them already use those preconstructed
dependencies. The only exception is the repository service which needs
to be able to execute the `gitaly-lfs-smudge` filter with a bunch of
environment variables derived from the configuration. We can eventually
encapsulate this functionality into its own dependency which can then be
injected, as well.

Remove the Gitaly configuration from all the other services.