git: Set up Git hooks via symlinks
When introducing the new temporary hooks directory created by the Git
command factory we had to resort to writing a wrapper script because
`gitaly-hooks` required the first argument to be set to the hook name
we're about to execute. This has been fixed now though, where it can
also use its zeroth argument to derive the hook name.

Refactor the code to not write a wrapper script anymore, but instead to
just create symlinks pointing at the `gitaly-hooks` binary. This avoids
having to spawn the shell just to execute the binary and thus reduces
one more point of friction.

Use of the temporary hooks directory is still guarded by a feature flag
which hasn't yet been rolled out, so we don't need a new feature flag
for this change.