rubyserver: Pass on BUNDLE_USER_CONFIG if set
On my local development environment I've set BUNDLE_USER_CONFIG to point
to ~/.config/bundle. This file contains:

    BUNDLE_PATH: "vendor/bundle"

So because BUNDLE_PATH is not set directly in the environment, it was
not picked up by Gitaly when it spawns gitaly-ruby.

With this change BUNDLE_USER_CONFIG is also propagated when
bundle-exec(1)'ing the Ruby sidecar, so it will be able to find the
gems.