Allow faster local builds
This applies Patrick's patch from:
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4192#note_810475985

The only changes from his original patch are:
* Rename "WITH_BUILD_ID" to "WITH_GNU_BUILD_ID" for clarity.
* Enable it by default.  Verified that disabling locally works.

The intent here is to allow developers to optionally speed up their
local builds by a few seconds, by using a single invocation of
"go build" rather than a separate invocation for each binary.

The only reason we need to run "go build" separately for each binary is
to provide each binary a unique value for GNU build-id.  We always want
that build-id to be present for binaries we may deploy or ship, but during
local builds, they are not needed unless doing local profiling.