Add option to disable metadata force creation in Praefect tests
Praefect contains a hack to create repository metadata when receiving
a request for a repository. This was put in place to enable us to use
the production code of Praefect in tests. That code requires database
records for repositories to be present. Praefect piggy backed on Gitaly's
tests for most of its RPC testing. Those tests broke with the new requirement
as they generally just place the repositories in the storage directly rather
than creating the repositories through the API which would allow Praefect
to create the metadata records. This testing hack potentially hides problems
in Praefect. As a first step towards removing it, this commit adds an option
to disable the testing hack in the test server. This allows for asserting a
given test no longer relies on the hack after being updated to create the
repositories through the API.