Add test helpers for creating repositories through the API
Praefect excepts repository creations to flow through it so it can
create the relevant metadata records. Gitalys tests are currently just
placing the test repositories directly on the disk which causes failures
in the tests when the tests are run with Praefect in front. This commit
adds two helpers to the gittest package to create test repositories through
the API, allowing Praefect to create the relevant metadata and better
matching the actual production scenario.