Expose a helper for checking whether Praefect is enabled in tests
There are some behavior differences between Praefect and Gitaly that
fail certain test assertions. The tests need to be adjusted for the
differences. To clearly mark the deviating behavior, this commit
exposes the IsPraefectEnabled helper that can be used to alter the
assertions of a test if Praefect is enabled. This also serves as a way
to document these behavior differences so they can be later amended.