Amend most operation service tests to create repos through the API
This commit amends most of the operations service's test to create the
test state through the API. This better matches the production scenario
and allows for Praefect to create the metadata it would create in real
setups. Most of the replacements are straight forward but some tests rely
on working trees. As Gitaly does not support creating worktrees through the
API, the worktrees are created after creating the repo through the API.

TestUserCommitFiles was simplified a bit and its branch was changed as new
repositories nowadays have main as their default branch.