Create test repository via API in TestUserApplyPatch
This commit creates the test repository via API in TestUserApplyPatch.
This change uncovered a small behavior difference in Praefect and Gitaly.
Now that Praefect is not using the metadata creation hack anymore, we can
see that Praefect's error message for non-existent repositories deviates
from Gitaly's. This is left as is since its been like this for a while
already with no apparent problems.