Disable metadata creation hack in operation service tests
Praefect piggybacks on Gitalys tests for testing the various RPCs.
The tests are simply run with Praefect in front of the Gitaly. This
worked fine before Praefect became stricter about consistency of the
operations and began requiring database records of the repositories.
The database records are updated when Praefect proxies the various
RPCs to the repository. Gitalys tests generally do not create the test
repositories via the API but place the directly in the storage. This
does not work with Praefect as Praefect won't see the changes and won't
be able to update its metadata. A hack was put in place in Praefect to
create the metadata in the tests when Praefect receives a request for
a repository. This is not ideal as it can hide bugs in Praefect.

As all of the relevant operation service tests have now been amended to
create their test repositories via the API, this commit disables the
metadata creation hack for operation service tests.