Setup test repository via API in repository service tests
This commit sets up the test repository via API in repository service
tests. This paves the way for disabling the Praefect's metadata creating
hack in the tests.

TestRemoveRepository_locking was updated to delete the lock after
the test so the repository creation helpers can delete the created
repository after the test.