Create repository via API in FetchRemote tests
This commits converts FetchRemote to also create repositories via
API. The tests were somewhat simplified by removing a method to copy
the test repository and instead initiate another copy of the test repo
using the existing helpers. As TestFetchRemote_sshCommand is overriding
the git command, it's not possible to create repositories via the API.
Thus, the test is skipped when Praefect is enabled.