Disable metadata force creation in repository service tests
This commit disables metadata force creation in repository service
tests when it is ran with Praefect. This ensures any problems in
Praefect are not hidden by the hack. Disabling the hack surfaces few
behavior deviations with differing error messages.

The RenameRepository tests do not work with Praefect without the
metadata creation hack. This is due to a bug in Praefect where the
rename still gets applied even if the Gitalys failed the rename.
As this is an existing bug, the test is skipped for now with a link
to the tracking issue.