repository: Extend test coverage for RepackFull & RepackIncremental
Make sure NotFound code and specific message is returned
in case requested repository doesn't exist on the disk.
In case repository is not provided (nil value) the proper
error should be returned. Because of the type conversion
from (*Repository)(nil) to repository.GitRepo interface
the variable 'repo' was not equal to nil as it's type
was set as *Repository, however the referencing value is nil.