operations: Respect timezone of committer in UserApplyPatch
When applying a patch via UserApplyPatch we use the user's credentials
as committer information. While we also use the user's timestamp if
passed to us, we fail to respect the user's timezone as configured in
their settings.

Fix this by using `dateFromProto()`, which knows to extract the timezone
of the user. Introduce a new feature flag which guards the new code
given that this is changes the current behaviour and thus leads to
different commit IDs.

Changelog: fixed